Some TestObjects defined to flex the generation correctness.  Just the normal members tested so far

+++ Noggolloquy.Generation/Fields/TypicalTypeGeneration/PercentType.cs
@@ -3,7 +3,7 @@ using System;
 
 namespace Noggolloquy.Generation
 {
-    public class PercentType : FloatType
+    public class PercentType : DoubleType
     {
         public override Type Type => typeof(Percent);
     }
